fix(uci): parse every `go` argument at upstream's own width, and its perft
Five divergences in the shell, all found by driving both engines with the
same scripts and diffing.

PERFT ZERO IS NOT A PERFT. Upstream dispatches on `if (limits.perft)`, so
`go perft 0` is an ordinary search with no limit at all. rfish read the
token's mere presence as "this is a perft" and printed a depth-zero divide
of every root move.

PERFT WITH NO ARGUMENT. Upstream's `is >> limits.perft` sets failbit and
terminates. rfish defaulted the depth to 1 and answered, so `go perft` and
`go perft zzz` produced a divide where upstream reports a critical error.
`perft` now goes through the same validated parse as every other key.

THE WIDTHS ARE THE ACCEPT/REJECT BOUNDARY. `is >> x` fails when the text
does not fit the FIELD's type, and upstream turns failbit into a critical
error, so the declared width is observable from a GUI. Everything was
parsed as `i64` here, which is too wide for the five `int` fields and too
narrow for `nodes`:

  go depth 3000000000               upstream: CRITICAL ERROR   rfish: searched
  go mate 3000000000                upstream: CRITICAL ERROR   rfish: searched
  go movestogo 3000000000 ...       upstream: CRITICAL ERROR   rfish: searched
  go nodes 18446744073709551615     upstream: searched         rfish: ERROR

Each key now parses at upstream's type: `TimePoint` (i64) for the clocks,
`int` for movestogo/depth/mate/perft, `u64` for nodes -- the last read the
way a C++ stream reads one, where a leading minus is accepted and the
magnitude wraps, so `go nodes -1` is a budget of u64::MAX on both sides.

PERFT DEPTH IS SIGNED. Upstream's `Depth` is `int` and its root test is
`depth <= 1`, so `go perft -1` counts one node per legal move. Carrying the
depth unsigned turns that into a perft four billion deep -- a stack overflow
rather than an answer -- so `perft_divide` now takes an `i32`.

`eval` LOST A LINE. Upstream prints `sync_cout << "\n" << Eval::trace(...)`,
so the blank line belongs to the COMMAND and is printed whether or not the
trace has anything to say. The two early returns -- in check, and no network
-- omitted it, so `eval` on a position in check came out one line shorter
here than upstream prints it.

BENCH NAMED THE WRONG POSITION. Upstream prints `engine.fen()`, the position
as it will be searched; rfish printed the FEN the entry was written with.
Twelve of the fifty-one bench entries carry moves, and for each of those the
`Position:` line named a position that is not the one the node count below
it belongs to.

Evidence (exit 0): 14 of 14 A/B cases against the pinned oracle agree, all of
which diverged before; golden-audit 13 agree, 0 differ; signature 2508687;
parity, no gate skipped. `tools/cases/golimits.uci` pins the widths, the
perft dispatch, the `eval` block and the error text.

+++ crates/rfish-engine/src/board/movegen.rs
@@ -464,12 +464,16 @@ pub fn perft(pos: &mut Position, depth: u32) -> u64 {
 
 /// Perft with the per-move breakdown UCI's `go perft` prints.
 #[must_use]
-pub fn perft_divide(pos: &mut Position, depth: u32) -> (Vec<(Move, u64)>, u64) {
+pub fn perft_divide(pos: &mut Position, depth: i32) -> (Vec<(Move, u64)>, u64) {
     let mut out = Vec::new();
     let mut total = 0;
     for m in generate_legal(pos).as_slice() {
         pos.do_move(*m);
-        let n = if depth <= 1 { 1 } else { perft(pos, depth - 1) };
+        // SIGNED, because upstream's `Depth` is `int` and its root test is `depth <= 1` --
+        // so `go perft -1` counts one node per legal move there. Taking the depth unsigned
+        // turned that into a perft four billion deep, which is a stack overflow rather than
+        // an answer.
+        let n = if depth <= 1 { 1 } else { perft(pos, (depth - 1) as u32) };
         pos.undo_move(*m);
         out.push((*m, n));
         total += n;
